astrocyte-services-py/astrocyte-rest/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astrocyte-rest"
version = "0.1.0"
description = "Reference HTTP server for the Astrocytes memory framework (REST over astrocyte-py)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "astrocyte==0.1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]
pgvector = ["astrocyte-pgvector"]

# `astrocyte` is not published to PyPI yet; resolve it from the sibling package when using uv.
[tool.uv.sources]
astrocyte = { path = "../../astrocyte-py", editable = true }
astrocyte-pgvector = { path = "../astrocyte-pgvector", editable = true }

[project.scripts]
astrocyte-rest = "astrocyte_rest.main:main"

[tool.hatch.build.targets.wheel]
packages = ["astrocyte_rest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
